+++ GEMstack/knowledge/defaults/current.yaml
@@ -39,7 +39,7 @@ simulator:
     real_time_multiplier: 1.0    # make the simulator run faster than real time by making this > 1
 
 pedestrian_detection:
-    model: !relative_path "../detection/yolov8n.pt"   #the !relative_path directive finds a file in a relative location to this file
+    model: !relative_path "../../knowledge/detection/yolov8n.pt"   #the !relative_path directive finds a file in a relative location to this file
     rate: 5.0
 
 longitudinal_planning:
