+++ GEMstack/onboard/perception/pedestrian_detection.py
@@ -6,8 +6,6 @@ import cv2
 from typing import Dict
 from ...utils import settings
 
-MODEL = settings.get('pedestrian_detection.model')
-
 
 def box_to_fake_agent(box):
     """Creates a fake agent state from an (x,y,w,h) bounding box.
@@ -26,7 +24,8 @@ class PedestrianDetector2D(Component):
 
     def __init__(self, vehicle_interface: GEMInterface):
         self.vehicle_interface = vehicle_interface
-        self.detector = YOLO(MODEL)
+        self.detector = YOLO(settings.get('pedestrian_detection.model'))
+        # self.detector = YOLO('./GEMstack/knowledge/detection/yolov8n.pt',verbose=False)
         self.last_person_boxes = []
 
     def rate(self):
