+++ GEMstack/onboard/planning/longitudinal_planning.py
@@ -1,8 +1,13 @@
 from typing import List
 from ..component import Component
 from ...state import AllState, VehicleState, EntityRelation, EntityRelationEnum, Path, Trajectory, Route, ObjectFrameEnum
-from ...utils import serialization
-from ...mathutils.transforms import vector_madd
+from ...utils import serialization, settings
+from ...mathutils.transforms import vector_madd, vector_dist
+
+
+DELTA_TIME = settings.get('longitudinal_planning.delta_time')
+# DELTA_TIME = 0.05
+
 
 def longitudinal_plan(path : Path, acceleration : float, deceleration : float, max_speed : float, current_speed : float) -> Trajectory:
     """Generates a longitudinal trajectory for a path with a
@@ -14,81 +19,89 @@ def longitudinal_plan(path : Path, acceleration : float, deceleration : float, m
        decelerate with accel = -deceleration until velocity goes to 0.
     """
     path_normalized = path.arc_length_parameterize()
-    #TODO: actually do something to points and times
-    times = [0]
-    points = [path_normalized.points[0]]
-    currrent_position = points[0][0]
-    remaining_distance = path_normalized.points[-1][0] - path_normalized.points[0][0]
-    braking_distance = (current_speed ** 2) / (2 * deceleration)
 
-    while times[-1] <= 10 and braking_distance < remaining_distance and current_speed < max_speed:
-        current_speed += acceleration * 0.05
-        current_speed = min(max_speed, current_speed)
+    start_point = path_normalized.points[0]
+    end_point = path_normalized.points[-1]
+    points = [start_point]
+    times = [0.0]
 
-        displacement = current_speed * 0.05
-        currrent_position += displacement
-        remaining_distance -= displacement
-        braking_distance = (current_speed ** 2) / (2 * deceleration)
+    current_point = points[0]
+    current_time = times[0]
 
-        times.append(times[-1] + 0.05)
-        points.append((currrent_position,0))
+    # points = [p for p in path_normalized.points]
+    # times = [t for t in path_normalized.times]
 
-    while times[-1] <= 10 and braking_distance < remaining_distance:
+    brake = False
 
-        displacement = current_speed * 0.05
-        currrent_position += displacement
-        remaining_distance -= displacement
-        braking_distance = (current_speed ** 2) / (2 * deceleration)
-        
-        times.append(times[-1] + 0.05)
-        points.append((currrent_position,0))
-    
-    while times[-1] <= 10 and current_speed > 0:
-        current_speed -= deceleration * 0.05
-        current_speed = max(0, current_speed)
+    while current_point[0] < end_point[0] and abs(end_point[0] - current_point[0]) > 0.0001:
+        current_time += DELTA_TIME
 
-        displacement = current_speed * 0.05
-        currrent_position += displacement
+        if not brake:
+            stopping_distance = (current_speed ** 2) / (2*deceleration)
 
-        times.append(times[-1] + 0.05)
-        points.append((currrent_position,0))
-        
+            distance_to_end_point = vector_dist(current_point, end_point)
+            # distance_to_end_point = round(distance_to_end_point, 2)
+
+            # print("distance_to_end_point:", distance_to_end_point)
+            if distance_to_end_point <= stopping_distance:
+                brake = True
+
+            else:
+                if abs(max_speed - current_speed < 0.0001):
+                    # stay on max speed
+                    current_speed = max_speed
+                    current_point = (current_point[0] + current_speed * DELTA_TIME, 0)
+
+                else:
+                    # we need to accelerate to achieve max speed
+                    current_x = current_point[0] + current_speed * DELTA_TIME + 1 / 2 * acceleration * (DELTA_TIME ** 2)
+                    current_point = (current_x, 0)
+
+                    current_speed += acceleration * DELTA_TIME
+
+                points.append(current_point)
+                times.append(current_time)
+
+        else:
+            # Case that we need to brake
+            if current_speed > 0:
+                current_x = current_point[0] + current_speed * DELTA_TIME - 0.5 * deceleration * (DELTA_TIME ** 2)
+                current_point = (current_x, 0)
+                current_speed -= deceleration * DELTA_TIME
+
+                points.append(current_point)
+                times.append(current_time)
+
+            # Case that we finished our brake since current speed <= 0
+            else:
+                break
+
     trajectory = Trajectory(path.frame,points,times)
     return trajectory
 
 
-def longitudinal_brake(path : Path, deceleration : float, current_speed : float) -> Trajectory:
+def longitudinal_brake(path : Path, deceleration : float, current_speed: float) -> Trajectory:
     """Generates a longitudinal trajectory for braking along a path."""
     path_normalized = path.arc_length_parameterize()
-    #TODO: actually do something to points and times
+    # TODO: actually do something to points and times
     times = [0]
     end_time = path_normalized.times[-1]
     points = [path_normalized.points[0]]
-    currrent_position = points[0][0]
+    current_position = points[0][0]
     remaining_distance = path_normalized.points[-1][0] - path_normalized.points[0][0]
     braking_distance = (current_speed ** 2) / (2 * deceleration)
 
     if current_speed == 0:
         while times[-1] <= end_time:
             times.append(times[-1] + 0.05)
-            points.append((currrent_position,0)) 
-
-    # while times[-1] <= 10 and braking_distance < remaining_distance:
-
-    #     displacement = current_speed * 0.05
-    #     currrent_position += displacement
-    #     remaining_distance -= displacement
-    #     braking_distance = (current_speed ** 2) / (2 * deceleration)
-        
-    #     times.append(times[-1] + 0.05)
-    #     points.append((currrent_position,0))
+            points.append((current_position, 0))
     
     while times[-1] <= 10 and current_speed > 0:
         current_speed -= deceleration * 0.05
         current_speed = max(0, current_speed)
 
         displacement = current_speed * 0.05
-        currrent_position += displacement
+        current_position += displacement
 
         times.append(times[-1] + 0.05)
         points.append((currrent_position,0))
@@ -143,7 +156,7 @@ class YieldTrajectoryPlanner(Component):
         #parse the relations indicated
         should_brake = False
         for r in state.relations:
-            if r.type == EntityRelationEnum.YIELD and r.obj1 == '':
+            if r.type == EntityRelationEnum.YIELDING and r.obj1 == '':
                 #yielding to something, brake
                 should_brake = True
         should_accelerate = (not should_brake and curr_v < self.desired_speed)
